requirements.txt
Django==1.7
distribute==0.6.34
dj-database-url==0.2.1
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
httplib2
simplejson
six
python-instagram